FREEI-3453: add sangomartapi_conference entry as an asterisk manager
We only need the manager entry for the conference app for now, but more will be added.  The rtapi apps will be setting listening for events through these accounts.  These manager entries should not be modified by admins, so they will not show up for editing in the Asterisk Managers gui.